relocated code from wrong slice(HeadersSlice.jsx) to headersSlice.js
